megano_backend/requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
diploma-frontend==0.6
Django==4.2.5
django-request-logging==0.7.5
djangorestframework==3.14.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==4.0.0
pytz==2023.3.post1
sqlparse==0.4.4
